Add limited support for diacritics
Use a wrapper for characters that allows to specify an additional diacritic. Diacritics do not need any
additional character definitions, but are drawn in code to save space.

Currently, only a diaeresis/umlaut is supported. While the diacritic will be drawn for any character, the
result may be unpleasant for characters that extend above the x-height.