Improve character spacing in high-resolution graphics
Instead of using only the minimum distance, the new algorithm tries to
use the average visual distance of the characters.

The minimum distance is still calculated first as a baseline, then the
algorithm approximates the average distance from all rows where the
actual distance is up to 3 pixels more than the minimum. Finally, the
spacing is limited based on both the character rect and the actual
resulting smallest pixel distance.